==> 7_final_info¬gelolocalizaciones.ipynb <==
import pandas as pd
import numpy as np

#pip install geopy

from geopy.geocoders import Nominatim
import time
import math

df=pd.read_csv('./6_final_info.csv')

df.drop(columns=['Unnamed: 0'],inplace=True)

df

#elimino los índices de las 4 ciudades que no tienen geolocalización
df.drop(df.index[647],axis=0,inplace=True)
df.drop(df.index[1218],axis=0,inplace=True)
df.drop(df.index[1309],axis=0,inplace=True)

df.reset_index()

df.drop(df.index[1218],axis=0,inplace=True)

df.reset_index()

#df[df['City']=='Mcmurdo Station'].index()

geo=Nominatim(user_agent='User')
latitude=[]
longitude=[]
for i in df.Ciudad_Country:
    loc=geo.geocode(i)    
    latitude.append(loc.latitude)
    longitude.append(loc.longitude)

print(latitude,longitude)

latitude=[38.97039795, 38.7077507, 25.7741728, 27.9580004, 47.4979937, 30.2711286, 41.1494512, -34.6075682, 41.6934591, 18.7882778, 42.6978634, 52.2319581, 13.7525438, 44.8178131, 41.0091982, 7.9366015, 32.751748750000004, 9.734950300000001, 16.068, 19.4326296, 50.0874654, -33.928992, 37.5666791, 1.357107, 43.6534817, 45.7538355, 52.5170365, 59.3251172, 39.4697065, 43.5116383, 3.1516964, -23.5506507, -7.4299703, 50.0619474, 12.9767936, 5.4065013, 33.7489924, 51.1263106, 20.6308643, 45.8131847, 44.4361414, 40.1776245, 36.9009641, -37.8142176, 21.2890997, -36.852095, 46.769379, 28.2935785, -33.8698439, 46.0500268, -27.5973002, 49.2608724, 40.4167047, 6.2443382, 25.0375198, 45.2551338, 51.5073219, 36.1672559, 45.6523093, 34.0536909, -8.5068977, 48.1371079, 55.6867243, 37.9839412, 41.3828939, -22.9110137, 37.1489538, 40.7127281, 59.4372155, 10.7715512, 39.7392364, 48.2083537, 21.1617854, 38.8950368, 42.3602534, 42.4248515, 37.7790262, -41.2887953, 52.3727598, 27.9477595, 40.6403167, 40.2111931, 53.550341, 22.2793278, 45.5202471, 41.8344086, 28.40037685, 14.5948914, 54.6870458, 42.1418541, 37.19420955, 51.0493286, 46.8183175, 41.8755616, -26.205, 7.0648306, 32.7762719, 30.544722, 20.9142238, 42.6502473, -0.2201641, 44.1767161, 20.2152358, 15.8693331, -12.0621065, 32.2228765, 49.1922443, 53.42968095, 42.4057219, 41.5510583, 37.808527350000006, 47.0245117, 23.135305, 27.3989266, 37.3886303, 10.2931062, 45.3969506, 12.1066695, 20.9876996, 31.2322758, 20.5927744, 44.9772995, 16.7967129, 44.9417468, 54.36119285, 56.9493977, 33.38034745, 47.6038321, 41.6509502, 48.4283182, 53.535411, 50.938361, 48.8588897, 50.8465573, 8.0610907, 19.0759899, 19.8887438, 40.7596198, 43.2073873, 13.3617562, 43.8519774, 52.4006632, 41.9960924, 41.3281482, -31.9558964, 40.640496, 48.1516988, -43.530955, 15.8798137, 43.2363924, -30.0324999, 51.3406321, -18.1415884, 10.2153093, 23.1301964, 14.9162811, 50.2137321, 35.6828387, 32.7174202, 11.568271, -0.62881505, -1.2832533, 11.083365350000001, 28.5421109, 20.6407176, 43.3211301, 51.772806599999996, 9.9051382, 50.1106444, 51.2254018, 50.6365654, 29.9759983, 57.1482429, 49.8955367, 43.6858146, 53.4071991, -8.4982478, 51.218194499999996, 58.3801207, -45.0321923, -25.2800459, 17.360589, -33.9617051, 34.7743988, 46.8836527, 54.9738474, 45.5548793, -25.7459277, 51.0460954, -2.8971725, 31.6258257, 25.6802019, 35.9127609, 52.131802, 34.6198813, 50.44876, 39.100105, 22.2973142, 36.0625843, 5.8247628, 20.0462328, 29.4246002, 38.1112268, 31.9515694, 6.112331, -34.9281805, 14.5589875, -8.6908357, 43.6044622, 5.9780066, 35.1799528, 26.2043691, 22.8015194, 45.3271752, 17.0, 1.5574127, 51.9244424, 43.1220119, 41.2230048, 23.7952809, 39.9622601, 12.9366674, 64.145981, 37.0162727, 43.541224400000004, -27.4689682, 33.88922645, 47.2186371, 55.6052931, 48.7784485, 16.4119945, 7.55984195, 44.3190159, 15.3503192, -25.1076701, 53.4794892, -8.6212371, 27.708317, 12.5672395, 44.1168594, 44.7720845, 49.1913033, 52.6361398, -16.9206657, 16.5735957, 36.550238, 43.6112422, 9.931308, 19.7073734, -34.9206797, 52.4796992, 34.6852901, 29.6519684, 20.6720375, 49.453872, 44.811349, 49.0521162, 39.7683331, 46.877229, 53.7435722, 18.4801972, -45.8740984, 12.3051828, 27.5236998, 30.0860459, 26.2122345, 40.4416941, 39.1014537, 33.6251241, 17.9712148, 12.2431693, 47.1615341, 19.7589517, 50.776351, -32.9595004, 39.613432, 22.68394485, -13.5170887, 37.270973, 35.4729886, 35.777103, 40.4864444, 55.17108, 30.3118769, 51.2211097, 45.94780155, 57.7072326, 33.4484367, 44.2337861, 38.3436365, 19.3580871, -12.9822499, 20.9670759, 44.648618, 37.9747645, 33.2336952, 35.2272086, 53.2190652, 48.23251, 25.9140256, 35.2072185, 32.3140354, 42.4415238, 37.5385087, 51.3582945, 44.841225, -54.806115899999995, 43.8975558, 34.7465071, 47.0708678, 39.7284945, 43.3435855, 42.498275, -16.3988667, -13.8343691, 30.6943566, 49.9655521, -3.7304512, 36.7213028, 42.8501191, 23.3700501, 37.0628317, 36.9863599, 29.7589382, 6.8137426, -7.2459717, 15.2241917, 50.9281717, 43.451291, 22.6203348, 22.5445741, -8.0584933, 36.0726355, 26.1947962, 35.3738712, 32.3669656, 47.3744489, 22.1757605, 49.0068705, 36.1556805, 34.6086854, 10.4195841, 35.7803977, 21.0294498, 36.3302284, 38.4224548, 41.6529143, 43.5460516, 26.3013982, 11.9477338, 39.9207886, 60.1674881, 45.7578137, 44.9391565, 46.097995, 38.7893909, 29.8988618, 31.8122623, -22.910832, 41.504879, 28.209538, 30.8327022, 35.9423833, 53.6466645, 32.4610708, 36.1622767, 50.013064, 43.0481221, 38.5810606, 41.1035786, 33.4709714, 32.3512601, 32.5010188, 51.47091115, 42.3315509, 11.9402416, 53.0758196, -6.9903988, 14.5567949, -10.3333333, 25.3356491, 8.5241122, 46.808327, 18.4724603, 28.8026443, 43.157285, 36.5436517, 37.8044557, 39.4667025, 47.5048851, 41.0799898, 30.1086537, 37.6922361, 6.3700386, 30.6598628, 38.7060457, 25.2653471, 48.7172272, 41.9818469, 36.06523, 0.5262455, -25.4295963, 25.4381302, 60.2047672, 36.5277607, -34.9058916, -42.8825088, 37.7742152, 36.7394421, 42.317099, -20.1624522, 48.2765903, 51.897077, 18.521428, 37.1641544, 22.33164705, 42.8620287, 43.6166163, 35.7688425, 33.5635206, 34.769536, -5.805398, 40.03813, 45.5031824, 34.0346534, 4.6533326, 22.9912348, 29.5541432, 6.9387469, 50.3219015, 42.444553, 42.0979745, 23.1146515, 25.7561647, 23.0205969, -22.5776104, 48.1113387, -1.45056, -8.334584, 44.40726, 9.9261153, 52.080985600000005, 2.3293744, 42.5814314, 43.7159395, 4.5986817, -34.4243941, 24.8843019, 51.0538286, 1.503555, 37.9577016, 40.6938609, 54.8982139, 39.08854049999999, 44.4938203, -5.1342962, 44.5126379, -33.4377756, 51.380926349999996, 34.5361067, 41.4086874, 45.7874957, -17.831773, 31.2237434, 30.3321838, 36.2516475, 35.2521996, 45.3573429, 42.8142432, 45.1875602, 34.6634288, 35.6892523, 0.3177137, 43.1579812, 30.247601, 4.8895453, 34.5793131, 32.0852997, 55.9533456, 43.0161693, -31.4173391, 12.054616800000002, 30.2265949, 37.104153, 52.0191005, -22.90556, 32.5007031, 39.063956, -8.8272699, -6.1753942, 40.1936892, -19.9227318, 34.851354, 9.3040623, 16.4639321, 52.6285694, 22.5816619, 46.2087066, 10.9428763, 32.844017, 26.1908241, 26.2235041, 6.3330586, 60.4517531, 52.23433665, 47.9078244, -18.9100122, 39.1670396, 19.439884550000002, 39.7686055, 35.8272257, -8.35760475, 44.7319094, 26.1223084, 39.9527237, 17.5211917, 40.335345, -4.2694407, 55.7121643, 53.3497645, 19.1943294, 36.1867442, 28.8638424, 26.1805978, 33.8439408, 31.3151708, 31.1171441, 35.5120831, 62.2393431, 31.098207, 55.8609825, 36.508976, 41.3082138, 39.94918695, 29.0948207, -37.9977225, -29.902631, 42.2885656, 34.0007493, 51.5142273, 35.1594647, 27.1752554, 41.6819935, 36.17262995, 40.3754434, 31.199004, 6.0328139, 7.3777462, 33.570499, 37.3361905, -1.950851, 51.1282205, 39.7589478, 33.5524785, 43.4652699, 28.4157769, -29.861825, 42.2799929, -17.333333, 33.5675527, 42.4184296, -20.4640173, -8.5837726, 20.28105765, 26.2159066, -20.2140657, 37.2633325, 25.702096, 43.074761, 40.1075089, 36.8968052, 52.1518157, 32.6405377, 35.6009498, 8.479004, 46.601557, 40.1827657, 22.8938884, 41.1257843, -11.6642316, 16.6835469, 38.2493581, -27.0988126, 65.0118734, 42.6638771, 22.0000001, 36.9903199, 40.8337963, 35.8713, 40.5863563, 51.7520131, 34.7494749, 35.3872218, 38.2542376, 26.2967719, -44.6941691, 42.8765615, 28.106471, 42.2681569, 52.5089759, -2.5295265, 40.6178915, 29.1871986, 44.0234387, 40.4233142, 15.3518378, 37.4946568, 52.3657098, 31.6343083, 27.9511188, 43.7698712, 30.1600827, 11.8948389, 40.6022059, 26.2037247, -23.1860383, 51.842574850000005, 40.4797828, 36.3134398, 41.6612561, -23.5003451, 8.034499199999999, 39.1404477, 9.0643305, 6.8393788, 10.5256264, 27.4989278, 46.8087794, -21.1343401, 21.19903535, 34.9236095, 23.6871297, 40.1164841, 47.41066, -16.4955455, 11.8763836, 37.7385507, 19.2716301, 29.03970805, 41.5910323, 30.5955289, 53.2744122, 39.2908816, -6.7137044, 41.7571701, 40.5092961, -17.3936114, 38.8339578, 41.4996562, 7.7382813, 42.1294712, 33.8829996, 43.8130735, 34.9531295, 33.9566098, 36.3967623, 33.1154683, 33.9533546, 22.9372087, 3.5896654, -16.680882, 34.0747444, 7.2930922, 53.7974185, 46.2062319, 43.17389195, 49.694285, 30.2240897, 34.5978596, 30.4211144, 16.5087586, -0.19299, 47.7981346, 37.6390972, 43.1009031, 35.021041, 37.7986184, 44.2613967, -15.4164124, 57.0462626, 41.8239891, 30.6743643, 34.3952765, 11.10684475, -4.05052, 17.7231276, 39.7990175, 32.182598, 32.5369356, 13.4410165, 38.6264256, 11.2450558, 40.8358846, 6.487679099999999, 23.356377, 50.9025349, 40.6549157, 42.9995032, 34.4263886, 21.2379469, 13.4727446, 24.8755492, 47.0451022, 27.9946969, 37.4137536, 4.7676576, 39.8454163, 31.3869323, 11.127834, 22.5414185, 42.9632405, 50.8036831, -16.5048228, 41.5235808, 46.1030443, 49.163877, 49.7596208, 48.7544012, 28.3754144, 37.08418, 28.6581723, 8.8879509, 41.9758872, 38.7566055, 18.833333, 45.6496485, 34.729847, 20.0112475, 21.2321047, 46.2017559, 42.5069391, 9.5170602, 35.1748976, 28.4963633, 29.6060218, 35.9556924, 23.2790444, 28.2996183, -20.3200917, 52.2055314, 32.0615652, 43.5854823, 34.022405, -6.9344694, 22.5213807, 45.5616075, 52.9534193, 52.5150949, 27.7237996, 37.870009, 36.0638034, 42.7337712, 48.584614, 29.5647398, -7.1215981, 44.3893113, 32.2990021, -6.1258942, 42.2713945, 36.7753606, 43.8014053, 43.3709703, 39.906217, 44.4761601, 7.1172633, 53.9590555, 5.947822, 41.3123363, 40.9965446, 52.3744779, 29.3093002, 32.5221828, 52.0799838, 37.8725487, 22.5000001, 51.5887845, 1.2146013, 10.6572678, 25.6093239, -21.7558619, 37.1734995, 36.3320197, 35.7032751, 59.9133301, 45.806691349999994, 33.7475203, 52.4081812, 51.456659, 45.92956455, 20.4317585, 31.549333, 50.7194784, 56.7291997, 31.666960500000002, 50.879202, -2.2895167, -23.6463741, 38.9181714, 35.6876096, 49.4093582, 49.872775, 13.6989939, 36.8529841, 10.50614075, 29.8622194, 31.3016935, 35.1851045, 24.5164639, 27.7951482, 21.19177615, -23.1887668, 66.4976214, 12.5268736, -8.5536809, -29.310054, 40.2337289, 42.6511674, 23.1102636, 36.4072377, 28.6517178, 19.292545, 55.7127529, 14.6222328, 13.0836939, 47.9184676, -18.9188041, 38.4493315, 42.8184538, 12.3681873, 40.8228016, -34.787093049999996, -8.8285734, 42.03726, 39.1649125, 45.4077172, 29.000652549999998, 18.465299, 6.130419, 11.0018115, -2.9888297, 51.8784385, 6.698081, 47.2654296, 42.3192548, 31.8228094, 21.9812746, -6.3718272, -12.6076318, 48.0517429, -24.6581357, -5.8951725, 51.3154546, 25.238704, 31.5841689, 36.77101545, 38.246242, -4.1030531, 36.2974945, 42.6829765, 47.6571104, 42.9956397, 24.163162, 43.3224219, 31.4933074, 33.9137085, 10.9810761, 8.7342288, 22.273734, 32.5866826, 30.4459596, 50.8389149, 22.7203616, 31.8782976, 51.585618749999995, 39.049011, 35.0457219, 41.6749893, 24.4480637, -3.1316333, 5.320357, 21.2737079, 41.1670412, 43.4887907, 41.8933203, 46.9482713, 19.7540045, 19.1214072, 22.3053263, 50.7255794, 35.1987522, 41.8692972, 34.7333334, 29.2108147, 48.7081906, 41.5263603, -8.6831728, 36.617331, -33.0458456, 32.8795022, 31.9973662, 31.7819242, 30.0443879, 29.9992425, 25.4557366, -17.853135, 41.139981, 36.0127749, 41.764582, -26.2227778, 34.4221319, 39.9395724, 44.5384728, 31.5118281, 56.1496278, 47.3523672, 35.444991, 35.7947315, 4.8472017, 51.984257, 38.2677554, 32.2352666, 35.124945, 6.1461619, 4.0429408, 26.6059432, 45.8992348, 51.8995685, 27.2939333, 39.0167979, 47.3215806, 38.2544472, 60.3943055, 40.8088861, 37.9317766, 33.4942189, 38.0739964, 13.519054, 33.9814746, 27.7809047, 43.9470982, 26.640628, 40.4906216, 33.577752399999994, 26.8381, 26.5914524, -8.7724902, 52.2695736, 54.596391, 35.48001575, 38.8462236, -33.0244535, 41.0534302, 45.7988273, 64.175029, 39.1723908, 45.9917589, 61.4980214, 19.29348285, 32.896672, 43.9492493, 47.6571934, 11.2320988, 42.4966815, 32.665135, 27.833333, 50.1171903, 34.6553944, 43.0349931, 42.7656251, 33.3864655, -13.9875107, 49.8879177, 63.4305658, -28.0023731, 23.2584857, 36.19924, -35.2975906, 10.506098, 24.3284746, 30.4380832, 36.0620781, 38.1301968, 13.524834, 49.5978157, 11.1427624, 4.1779879, 29.9870753, -51.69487225, 20.8689257, 43.2961743, 43.061936, 10.2375144, 24.4801069, 30.2489634, 41.2587459, 51.2085526, 17.9640988, 0.5184021, 6.809107, 53.0775961, 45.6794293, 28.5811076, 42.5006243, 39.5261206, 32.6248622, 51.4816546, 40.0149856, 3.8689867, -15.9277296, 31.9827896, -6.8160837, -7.9771206, 27.9963899, 14.7421699, -8.6524973, 18.062922200000003, 35.3676472, 18.547327, 26.1003392, 24.7807798, 39.579155, 24.4538352, 21.1498134, 26.9154576, 44.0505054, 22.050466550000003, 45.4371908, 32.8406946, 51.4392648, 17.7395201, 34.3916058, 69.649208, 42.8867166, 40.6460921, 22.8193063, 10.6498095, 42.2625621, 30.8942995, 49.4404591, 9.0107934, 35.0841034, 37.0664467, 40.4590505, -25.966213, 38.4404925, 35.5570501, 43.5515198, 45.4641943, 10.3825318, 27.3365805, 34.6422939, 23.0216238, 45.4871723, 37.042257, 36.4071112, 13.0977832, -21.2074736, 34.3430507, 32.6200973, 34.4838502, 31.8457149, 18.556551, 35.1343368, 52.707755, 29.299328, 26.4609135, 33.4946353, 40.0919817, 52.1067449, 31.3536127, 38.0464066, 17.1527588, -7.9778383999999996, 31.9030543, 28.687547, -7.5692462, 18.9194213, -34.9632342, 38.9719384, 33.6306299, 32.7876012, 28.42826235, -8.5199633, 53.3806626, 44.0581728, -16.3332828, 23.1608938, 15.3389667, 26.7300936, -9.437797549999999, 33.1580933, 39.7034723, 37.4443293, 31.263235, 30.5951051, 12.61326555, 34.6932379, 33.4270275, 43.1672254, -28.6483333, 35.1433533, 30.9090157, 43.7009358, 33.6170092, 40.7896239, 61.2163129, 26.715364, 25.1737019, 39.2169525, 34.1976308, 39.0856735, 32.44645, 37.3893889, 41.1075276, 35.9170217, 40.0940758, 40.34833555, 17.1184569, 15.35, 40.63367, 40.924271, 24.5625566, 42.3636331, 37.5120997, 39.7459468, 30.421309, 38.4852037, 46.7729322, 39.415779, 37.4334738, 33.9516458, 38.974390099999994, 35.1335022, 34.2066591, 36.140807, 23.7861979, 44.0869329, 40.5508527, 45.4384958, 46.5927425, -4.3217055, 43.6610277, 33.0010496, 34.7487244, 43.2630018, 17.5000543, 37.4519671, 42.437072, 36.2231079, 32.0809263, 24.8546842, 19.877263, 36.343694049999996, 16.0000778, 11.2536716, 33.1958696, 33.1141426, 5.5571096, 33.2000368, 30.7334421, 44.5645659, 62.76712545, 37.496904, 49.1856637, 35.1139509, 26.4367824, 43.73844905, 4.6373012, 39.4862777, 37.9716991, 47.5581077, 29.101628050000002, -8.7278354, -12.46044, 24.471153, 43.4832523, 52.383705750000004, 34.6196539, 51.1777781, 25.0783456, 33.743764999999996, 30.1957677, 31.79592425, 49.6112768, 38.5425835, 44.5263107, 38.2325829, 53.2005936, 30.7474425, 39.5404679, 0.390002, 18.4257128, 26.12744855, 26.339806, 39.5182427, 43.419754, 46.8701049, 39.2501246, 38.0429742, 25.2856329, 35.9603948, 3.4517923, -17.7414972, 35.9957253, 27.4989086, 36.6507007, 38.2971367, 44.9391565, 32.010998, 45.639544, 6.5269033, 19.30037305, 44.766764249999994, 10.4541312, 38.8579735, 32.555664, 44.0206919, 2.0349312, 41.9602718, 25.3575436, 52.3713394, 58.42673005, 30.3674198, 36.1911624, 39.915463, 45.0677551, 33.2095614, 32.2956076, 18.0250713, 33.3024309, 32.6707877, 44.017763900000006, 46.5218269, 31.5656822, 23.61515, 36.7071591, 31.4649685, 46.0050102, 33.822986549999996, 24.638916, 33.74079965, 11.0248119, 18.341137, 10.6333401, 38.50935, 36.364519, 9.50139445, 19.7490157, 30.694088, 36.1023552, 46.6855231, 31.2457442, 32.0898005, 36.3496, 47.4250593, 33.6038766, 18.3403061, 38.7367448, 14.693425, 34.0123846, 27.8328352, 32.8191218, 39.6295951, 29.394644, 37.451434500000005, 42.5846773, 34.5260109, 36.819626, 12.1459907, 12.4345376, 36.583333, -54.283545, 33.6938118, 9.3348046, 37.9404648, 10.302834, 31.4220558, 30.1979793, 32.1288331, 14.1057433, 32.1525312, 10.0844025, 47.6144219, 41.903411, 30.4952371, 37.461928, 32.4935378, 4.9371143, 14.6367927, 15.5062156, 12.1472869, -9.4743301, 14.8954912, 41.9263991, 31.3231643, 15.5635972, 29.3797091, 34.7109786, 33.5130695, 11.8145966, 28.1450774, 4.3907153, 16.3217674, 32.3969935, 21.5810088, 21.420847, 6.328034, -22.2745264, 12.1191543, 50.4500336, 55.7504461, 49.841952, 53.9024716, 46.4843023, 48.7481718, 59.938732, 54.7261409, 49.9923181, 45.7684014, 63.3233807, 48.4680221, 47.9102734, 56.866557, 55.1598408, 51.530018, 54.991375, 56.839104, 54.710128, 54.3107593, 55.0951738, 47.2610441, 53.198627, 55.0282171, 58.324653850000004, 43.17199225, 55.7823547, 56.3264816]                                 

longitude=[-9.399580512067734, -9.1365919, -80.19362, -15.6062305376665, 19.0403594, -97.7436995, -8.6107884, -58.4370894, 44.8014495, 98.9858802, 23.3221789, 21.0067249, 100.4934734, 20.4568974, 28.9662187, 98.3529292, -16.981752048847362, 100.0305710845208, 108.212, -99.1331785, 14.4212535, 18.417396, 126.9782914, 103.8194992, -79.3839347, 21.2257474, 13.3888599, 18.0710935, -0.3763353, 16.4399659, 101.6942371, -46.6333824, 112.46309700455788, 19.9368564, 77.590082, 100.2559077, -84.3902644, 16.97819633051261, -87.0779503, 15.9771774, 26.1027202, 44.5126174, 30.6954846, 144.9631608, -157.717299, 174.7631803, 23.5899542, -16.621447121144122, 151.2082848, 14.5069289, -48.5496098, -123.113952, -3.7035825, -75.573553, 121.5636796, 19.8451756, -0.1276474, -115.148516, 25.6102746, -118.242766, 115.2622931, 11.5753822, 12.5700724, 23.7283052, 2.1774322, -43.2093727, -8.702467283736809, -74.0060152, 24.7453688, 106.6983801, -104.9848623, 16.3725042, -86.8510468, -77.0365427, -71.0582912, 18.7710437, -122.419906, 174.7772114, 4.8936041, -82.458444, 22.9352716, -8.4294632, 10.000654, 114.1628131, -122.6741949, 23.4841698, -14.004869713542156, 120.9782618, 25.2829111, 24.7499297, -8.585577395168343, 13.7381437, -71.2143518, -87.6244212, 28.049722, 125.6080623, -96.7968559, -9.709128, -100.743784, 18.0924947, -78.5123274, 28.6507598, -87.42953531439221, -97.0726428, -77.0365256, -110.9748477, 16.6113382, 14.592912670292268, -82.1853837, -8.4280045, -25.473085286479595, 28.8322923, -82.3589631, 33.6741556, -5.9953403, 123.9020773, -75.6851593, -68.9351307, -101.0, 121.4692071, -100.3902245, -93.2654692, 96.1609916, 26.0236504, 18.62860883362069, 24.1051846, 126.54825445912047, -122.3300624, 41.6360085, -123.3649533, -113.507996, 6.959974, 2.3200410217200766, 4.351697, 98.9161607, 72.8773928, 102.135898, -111.8867975, 27.9166653, 103.8590321, 18.3866868, 16.91973259178088, 21.4316495, 19.8184435, 115.8605801, -8.6537841, 17.1093063, 172.6366455, 108.3273939, 76.9457275, -51.2303767, 12.3747329, 178.4421662, 103.9880443, 113.2592945, -23.5095095, 19.00588775660632, 139.7594549, -117.1627728, 104.9224426, -90.3638752022324, 36.8172449, 120.91639641658887, -81.3790304, -105.220306, 21.8959232, 19.47849495849145, 126.05118878624401, 8.6820917, 6.7763137, 3.0635282, -90.0782127, -2.0928095, -97.1384584, -79.7599337, -2.99168, 116.0521267, 22.554677562145155, 26.72245, 168.661, -57.6343814, 78.4740613, 25.6207519, 32.4231586, 17.846663499449168, -1.6131572, 18.6953685, 28.1879101, -114.065465, -79.0043742, -7.9891608, -100.315258, 14.5020708, -106.660767, 135.490357, -104.61731, -94.5781416, 73.1942567, -94.1574328, -55.1703941, 110.1956502, -98.4951405, 13.3524434, 35.9239625, 125.1717145, 138.5999312, -90.7333258, 115.1641975, 1.4442469, 116.0728988, 129.0752365, -98.230082, 86.2029579, 14.4412309, -96.5, 110.3439862, 4.47775, -123.4131383, -111.9738429, 86.4309638, -83.0007065, 100.8864587, -21.9422367, -7.9351771, -96.73366465001021, 153.0234991, 35.50255852895232, -1.5541362, 13.0001566, 9.1800132, 120.5933729, 99.05728875374527, 23.7965614, 74.10178169795284, -64.3494964, -2.2451148, 115.126168, 85.3205817, 99.954586, 15.2353257, 17.1917651, -122.8491439, -1.1330789, 145.7721854, 120.4089899, -82.5594293, 3.8767337, 76.2674136, -155.08158, -57.9537638, -1.9026911, 33.0332657, -82.3249846, -103.338396, 11.077298, -91.4984941, -122.329479, -86.1583502, -96.789821, -0.3394758, -69.942111, 170.5035755, 76.6553609, -99.497352, -94.1018461, 127.6791452, -79.9900861, -84.5124602, 130.6180016, -76.7928128, 109.1898675, 27.5836142, 99.6734592, 6.083862, -60.6615415, 2.880353635614984, 91.78913897849154, -71.9785356, -79.9414313, -97.5170536, -5.803792, -86.1336351, -118.7949873, -95.4560512, 4.3997081, -66.65347069241668, 11.9670171, -112.074141, 28.6255665, -0.4881708, 98.439839, -38.4812772, -89.6237402, -63.5859487, -87.5558483, -8.500424800741666, -80.8430827, 6.5680077, -101.296273, -97.4890856, -101.8338246, -106.7798078, 19.2621081, -77.43428, 7.473296, -0.5800364, -68.3184972880496, -78.8635324, -92.2896267, 15.4382786, -121.8374777, 17.8076584, -92.3328743, -71.5369607, -171.7692793, -88.0430541, 16.9705651, -38.5217989, -4.4216366, -106.325138, 85.3250387, 37.3792617, 35.3252861, -95.3676974, -58.1624465, 112.7378266, 101.9313754, 11.5879359, -80.4927815, 120.3120375, 114.0545429, -34.8848193, -79.7919754, -98.1836216, -119.0194639, -86.3006485, 8.5410422, 113.5514142, 8.4034195, -95.9929113, -98.3903305, -75.5271224, -78.6390989, 105.8544441, -119.2920585, 27.1310699, -83.5378173, -80.2493276, -98.1624501, 121.9392632, 32.8540482, 24.9427473, 4.8320114, -123.033121, -64.80011, 0.1633256, -93.9288723, 119.9691539, -68.2001376, -73.9696822, 83.991402, -83.2784851, -77.7956949, -1.7822482, -84.9880449, -86.7742984, 22.016143457294312, -76.1474244, -121.493895, -80.6520161, -81.9748429, -95.3010624, -116.964662, -2.5790334025708415, -83.0466403, 108.4375758, 8.8071646, 110.4229104, 121.0211226, -53.2, 83.0076292, 76.9360573, -100.783739, -77.9217357, -96.9766308, -77.615214, 32.0019211, -122.2713563, -87.4139119, -111.29189, -85.1386015, 78.2916193, -97.3375448, 99.7928634, 104.0633717, -88.0949352, 55.2924914, 21.2496774, 44.1117811, -119.0167679, 101.4515727, -49.2712724, 81.8338005, 24.6568435, -87.3588703, -56.1913095, 147.3281233, -87.1133304, -119.7848307, -83.0353434, 57.5028044, -116.553247, -8.4654674, 73.8544541, -120.7678602, 103.85753892309174, -112.450627, -116.200886, -119.2470536, -101.879336, -92.2670941, -35.2080905, -76.3056686, -73.5698065, -5.0161926, -74.083652, 120.184982, -81.2207673, 79.8541134, 11.9178807, 27.208264534658166, 19.0954528, 114.4107202, 112.72964, 113.7457788, 17.0772739, -1.6800198, -48.4682453, 115.6409359, 8.9338624, 78.1140983, 5.12768396945229, 102.2880962, 18.2177461, 10.4018624, 101.0900236, 150.89385, 102.8324209, 3.7250121, 103.7495586, -121.2907796, -89.5891008, 23.9044817, -120.05035277301698, 11.3426327, 119.4124282, -88.0125794, -70.6504502, -2.37418971512966, -117.2911565, -75.6621294, -108.49607, 31.045686, -85.3933906, -81.655651, -118.852583, -92.6888936, -122.6067583, -73.9395687, 5.7357819, 135.5032721, 51.3896004, 32.5813539, -79.2441003, 121.31938451585901, 114.9417574, -118.1171108, 34.7818064, -3.1883749, -83.6900211, -64.183319, 102.33734259872213, -93.2173759, -113.5841313, 8.531007, -47.06083, -94.74049, -108.5507317, 13.2439512, 106.827183, -85.3865271, -43.9450948, -82.3984882, -82.12848154978464, 107.5863388, 1.2923049, 113.0761073, -119.119919, 108.2830253, -97.1430671, -97.6959794, 50.5822436, 5.6221058, 22.2670522, -0.9028072768185829, -97.0592028, 47.5255809, -86.5342881, 72.8803834000775, -94.8466322, -90.6949871, 116.08115279184318, -93.21772, -80.1433786, -75.1635262, 102.6680012, -75.9279495, 15.2712256, 12.563150889107241, -6.2602732, 72.9701779, -94.1288142, 78.80577833091104, 91.753943, 9.400138, -85.8552161, -97.727796, 24.0191544, 25.7458236, -97.3427847, -4.2488787, -4.88562, -72.9250518, 4.0529515568716175, -110.9692202, -57.5482703, -71.2520253, 18.8419505, -81.0343313, 7.4652789, 126.8515034, 78.0098161, -85.9766671, 27.919418145633188, 49.8326748, 29.894378, 80.214955, 3.8972497, -86.765783, -121.890583, 30.061507, 71.4306682, -84.1916069, 119.1075773, -80.5222961, 79.45264865437258, 31.009909, 118.9426131, -61.5, -7.618756606822727, -71.1061639, -54.6162947, 116.10685, 85.81140263163533, -98.3252932, -70.1524646, 127.0287472, 32.647186, -89.3837613, -89.3768539, -76.2602336, 4.4811088666204295, -85.4626848, -82.5540161, -13.26795, -120.510842, 29.0677305, -109.9200604, 16.8620293, 27.4826264, -92.61551841991157, -122.039966, -48.64454838500228, 25.4716809, 21.1640849, -102.5000001, -86.4436018, 111.6730788, 128.6018, -122.391675, -1.2578499, -77.4208221, -94.4248983, -85.759407, 73.0351433, 169.1364637, 74.6070079, -80.6371513, -83.7312291, 6.0943765, -44.2963942, -75.3786521, -82.1400923, -92.4630182, -104.7091322, 75.1379848, -120.8465941, 5.2210432, 74.8736788, 116.3530413, 11.2555757, -85.6545729, 8.5364136, -75.4712794, 78.1573628, -45.8866816, 5.838960628748229, -88.9939147, -82.3534728, -91.5299106, -47.4582864, 98.81788050252146, -121.6169108, 7.4892974, 81.8327629, 76.2132542, -82.5748194, 130.3734556, -175.201808, 81.3979545573657, 33.6236184, 86.9746587, -88.2430932, 40.101986, -68.1336229, 75.3737973, -121.4201395, 73.23586552517415, -13.636291608604138, -93.6046655, -96.3071042, -9.0490601, -76.610759, 108.5608483, -88.3147539, -88.9843937, -66.1568983, -104.825348, -81.6936813, 98.7709086, -80.0852695, 130.8749015, 125.3171216, -120.4358577, 116.7928916, -97.8791341, 129.7874339, -117.3961623, 89.2852741, 98.6738261, -49.2532691, 74.8204443, 80.6350768, -1.5437941, -90.6802594, 16.55652349457852, -112.851562, -92.0198427, 119.215813, -9.5830626, 80.6185102, -79.272301, 13.0464806, -120.9968782, -75.2326641, 135.7556075, -121.2117148, -88.4069744, 28.2824793, 9.9215263, -71.4128343, -96.3699632, 129.31544896587815, 76.10995510466662, 39.667169, 83.3012842, -89.6439575, -95.7893178, -93.6952689, -16.56275092072591, -90.1995853, 75.7754716, 14.2487679, 99.3024780130819, 116.6775918, -1.404189, 109.8344586, 125.9816054, -117.3008784, 81.6336833, 144.75201822544284, 118.6723605, -122.8950075, -80.6366144, -79.1422464, 7.0188527, -88.9524151, 120.9764683, 77.3415747, 88.35769124388872, -85.6678639, -1.075614, -68.1624337, -90.5770967, 11.129840331695068, -123.938122, 6.6441878, -122.478836, 121.3812013, -94.51323, 121.4163876, 76.5955013, -91.6704053, -90.42352, -98.0, 13.7772781, -86.5859011, 73.7902364, -86.73222812171971, 6.1466014, 1.5212467, -13.6998434, 33.3638568, 34.5145652, 52.5378041, -80.0053176, 116.0876485, -82.4522702, -40.3376682, 0.1186637, 118.7915079, 39.723109, -6.834543, 107.6049539, 113.3656141, -94.1642004, -1.1496461, 5.4768915, 106.9191408, 112.5452739, 120.3781372, -84.5553805, 7.7507127, 106.5478767, -34.882028, -79.6901736, -90.1847691, 23.5998111, -89.093966, 3.0601882, -91.2395429, -8.3959425, 116.3912757, -73.212906, -73.1156668, -1.0815361, 80.5482919, 69.2787079, 15.1405690365004, 9.7385532, 120.0703805, -93.7651944, 4.3113461, 32.4942749, -100.4949145, 4.7760237, -77.2782755, -61.5180173, 85.1235252, -41.3326895, -3.5995337, -94.1185366, -0.6492976, 10.7389701, 12.206315763116372, -116.9719684, -1.510477, -0.9696512, 6.934005265146571, -86.92027456597663, -97.1466695, -1.8767712, -111.3885221, 120.79080112776134, 4.7011675, -80.0983887437901, -70.3980033, 121.6282945, -105.9384569, 8.694724, 8.651177, -89.1914249, -75.9774183, 103.26205644863938, 121.6203873, 120.5810725, 136.8998438, 117.6424895, -97.3942053, 72.9544157804841, -46.884506, 25.7192101, -70.0356845, 125.5784093, 27.478222, -111.6587085, -73.754968, 112.99801241836545, -105.573284, 77.2219388, -99.6569007, 21.1350469, -90.5185188, 80.270186, 106.9177016, -48.2767837, -78.8688833, -1.6442556, -1.5270944, 114.880567, -68.43818677312292, 115.087937, -88.2810994, -119.7666496, 11.8734455, 77.76822875372174, -66.116666, 1.215829, 76.9628425, 104.756857, -0.4152837, -1.6230404, 11.3927685, -85.1824269, 117.2218033, 96.082375, -77.8093378, 15.7411039, -122.1768209, 25.9088474, 22.408589, 9.4924096, 110.1750728, 130.543387, 3.509589562138503, 21.7350847, -81.0481341, 59.6059232, -89.0226793, -117.2613936, -71.4547891, 120.6478282, -1.9838889, 120.3074357, -98.4933873, -74.7988827, 76.7257086, 113.5721327, 117.0130019, -91.18738, -0.12693017575934035, 75.8681996, 120.5514595, 5.066061908648313, -95.677556, -85.3094883, 123.458674, 118.3856331, -59.9825041, -4.016107, 110.3548318, -73.2048348, -112.03628, 12.4829321, 7.4514512, 96.1344976, -70.6432141, 70.8028377, -3.5269497, -111.6518229, 123.924157, 36.7166667, -81.0228331, 44.5153353, -88.0840212, 115.2564626, 101.7761976, -71.6196749, -111.757352, -102.0779482, -95.6258199, 31.2357257, 120.576854, 119.0029034, 25.861429, -104.820246, -5.6048873, -72.6908547, 27.89, -119.7026673, 119.5894129, 10.935960870530739, -9.7620903, 10.2134046, 123.9164653, 139.636768, -107.21208446076524, 31.5951655, 5.9108573, 140.8691498, 76.326238, 7.901173504229375, 6.8019088, 9.706203, -81.9806771, 6.1288847, -2.0711559, -80.3503283, 125.7473609, 5.0414701, -104.609141, 5.3259192, -96.7077751, -121.696026, -111.9260184, 46.2961952, 144.83783877764978, 134.3649289, 120.6512652, 12.630768610516004, -81.8723084, -98.9472344, -117.18845420205307, 80.9346001, 127.9773062, 115.1723944, 6.236339604766698, -5.9301829, 44.360072742328676, -77.3063733, -71.5517636, -73.5387341, 126.5303997, -51.7355386, 127.4285128, 9.264881035005711, 23.7603118, -100.46459244066614, 13.1777923, 4.8059012, -117.4235106, -74.1950883, -96.4058782, -114.47603157249804, 78.166667, -122.9543022, 133.9194595, -87.922497, -71.4677032, -86.78247922885197, 33.768144, -119.4959025, 10.3951929, 153.4145987, 77.401989, 37.1637253, 149.1012676, -66.9146017, 109.4239676, -84.2809332, 103.8318566, -121.2724473, 2.109823, 11.0038064, -74.1169296, 73.5107387, 31.2118063, -57.84522966123792, -105.4407022, 5.3699525, 141.3542924, -67.5890216, 118.0853479, 120.2052342, -95.9383758, 3.226772, 102.6133707, 25.2057292, -5.273263, 4.9057311371764385, -111.044047, 77.3031364, -90.6647985, -119.8126581, -115.448325, -3.1791934, -105.270545, 11.5213344, -5.7160872, 120.8904588, 39.2803583, 112.6340291, 120.695345, -91.1559757, 115.2191175, -15.962904570040749, 24.4736079, -72.3395928, -80.399513, 110.4922502, -104.8769227, 54.3774014, 79.0820556, 75.8189817, -123.0950506, -159.55876775017003, 12.3345898, -83.6324022, 5.478633, -88.03365155356477, 132.4518156, 18.9543434, -78.8783922, -111.4979963, 108.3627211, -71.6443596, -71.8018877, 120.0851114, 1.0939658, 38.7612525, -106.650985, 27.484418184610497, 128.9038646, 32.56745, -122.7141049, 45.4425963, 7.0134418, 9.1896346, 7.8533226, -82.5308545, 50.8801184, 72.5797068, -122.80378, 43.0953954, 25.456663652707153, -59.6184184, -159.7708145, 108.9350839, -83.606568, -114.3224495, -102.367687, -68.3691611, 36.7496276, -2.7540658, -94.7945882, 80.3217588, -117.1473661, 113.2951122, 4.2736937, 118.4277477, -84.4970393, -89.0761718, 110.36722565020224, 35.1952255, 115.8540042, 110.828448, 102.4477333, -54.9439981, -95.2359496, 114.6423617, -79.9402728, 77.00270014657752, 179.1982548, -1.4702278, -121.3153096, -48.9525756, 79.9497702, 38.9326763, -12.9766773, 159.9624174786047, -117.3505966, 125.8888149, -122.1598465, 32.305505, 114.2999353, -7.984739136241295, 135.1943764, -117.6124179, 11.467181161339118, 153.6177778, 110.769063, 75.851601, 7.2683912, -117.9294401, -73.9598939, -149.894852, -80.0532942, 75.8574194, 9.1128513, -119.1803818, 117.1951073, -99.7475905, -122.0832101, 122.9902689, 139.4886305, 124.3977559, -74.6684230570877, -61.8448509, 44.2, 14.6026095, -98.338685, -81.7724368, -87.8447938, 122.1152599, -75.546589, -87.2169149, 106.2261926, -92.1251218, -77.4127562, 118.6689349, 131.2468368, 1.4197463178515335, -89.9668758, 117.278286, -5.3541295, 90.4026151, -103.2274481, -105.0668085, 10.9924122, -112.036277, 15.3125974, -70.2548596, 112.5292693, 113.6191483, -2.9350039, -88.2003115, -122.1779927, -76.54831458600927, -121.3877428, -81.0911768, 67.0207055, 75.3390241, 43.099715597302286, -61.7333373, -85.8709492, -117.3794834, -86.8473788, -0.2012376, -117.2425355, 76.7797143, -123.2620435, 6.168828066313526, -122.3330573, -2.1102277, 117.1600052, 50.1039991, 7.424224092532953, -75.569799, -75.0254256, 126.5527573, 7.5878261, -81.01059617608054, 115.1730049, 130.8410469, 39.6111216, -1.5592776, 4.643559696559635, 112.4477046, -115.5682504, -77.3383331, -116.22331524966393, 67.0172447, 35.21198075969497, 6.129799, 68.81521423589763, -109.056392, -122.636465, 5.7918548, 120.7510971, -119.7487235, 9.454001, -64.6232227, -80.25779068328761, -81.7786972, -119.988811, 87.319461, -113.995267, -119.951909, 114.5088385, 51.5264162, -83.9210261, -76.5324943, 168.3150163, 119.4041185, 113.1529127, 117.1140042, -122.2855293, -123.033121, 112.1163785, 9.278869369225038, 3.5774005, -81.37009366377936, 11.827938995640167, -84.0096613, 115.490696, -95.8640507, -88.5408574, 45.3419183, 129.96778069176605, 55.3882729, 4.5307624, 8.94694911492874, -89.0928155, 44.0094652, 127.5439794, 7.6824892, -87.5675258, -64.7827048, -63.0483073, 44.3787992, 51.6650002, 10.454430026192636, 6.6327025, 74.3141829, 58.5912467, 119.1557441, -100.4405094, 8.9520281, -116.46687017113196, 46.7160104, -116.36936983394352, 7.651853311586596, -64.932789, -85.4362722, 125.7555447, 6.60826, 99.99561927320003, -70.4128435, 111.2803074, 114.33633, 7.8585139, 34.7925181, 72.6782574, 127.3848, 9.3765878, 73.048136, -88.49537037472777, 125.4115285, -17.447938, 71.5787458, 112.9388078, 34.9983856, 118.1738712, 71.6638747, 25.392337493319413, -87.8212263, 69.1776838, 117.980675, -86.2746665, -86.8773498, 114.483333, -36.4946355, 73.0651511, -83.5720475, 58.3823487, -85.8385625, 73.0923253, 71.4724978, 20.0817204, -87.2040052, 74.1933745, 99.8258031, -122.192337, 12.4528527, 47.8090981, 121.4425255, 74.5411575, -52.3258307, -61.01582685063731, -88.0248937, -68.2740206, 147.1599504, -88.92572717284826, 8.7376029, 48.6792538, 32.5349123, 47.9735629, 137.7259431, 36.3095814, 42.8453061, 113.2384362, 18.5509126, -86.17868327530661, 119.4077008, 39.1653612, 39.826869, -10.797788, 166.442419, 15.0502758, 30.5241361, 37.6174943, 24.0315921, 27.5618225, 30.7322878, 24.5207477, 30.316229, 55.947499, 36.2310146, 39.0261044, 97.0979974, 35.0417711, 33.3917703, 53.2094166, 61.4025547, 46.034683, 73.371529, 60.60825, 20.5105838, 48.3642771, 36.611913, 39.72490856518773, 50.113987, 82.9234509, 81.52937237232149, 131.96425966543694, 49.1242266, 44.0051395]

df['Latitude']=[38.97039795, 38.7077507, 25.7741728, 27.9580004, 47.4979937, 30.2711286, 41.1494512, -34.6075682, 41.6934591, 18.7882778, 42.6978634, 52.2319581, 13.7525438, 44.8178131, 41.0091982, 7.9366015, 32.751748750000004, 9.734950300000001, 16.068, 19.4326296, 50.0874654, -33.928992, 37.5666791, 1.357107, 43.6534817, 45.7538355, 52.5170365, 59.3251172, 39.4697065, 43.5116383, 3.1516964, -23.5506507, -7.4299703, 50.0619474, 12.9767936, 5.4065013, 33.7489924, 51.1263106, 20.6308643, 45.8131847, 44.4361414, 40.1776245, 36.9009641, -37.8142176, 21.2890997, -36.852095, 46.769379, 28.2935785, -33.8698439, 46.0500268, -27.5973002, 49.2608724, 40.4167047, 6.2443382, 25.0375198, 45.2551338, 51.5073219, 36.1672559, 45.6523093, 34.0536909, -8.5068977, 48.1371079, 55.6867243, 37.9839412, 41.3828939, -22.9110137, 37.1489538, 40.7127281, 59.4372155, 10.7715512, 39.7392364, 48.2083537, 21.1617854, 38.8950368, 42.3602534, 42.4248515, 37.7790262, -41.2887953, 52.3727598, 27.9477595, 40.6403167, 40.2111931, 53.550341, 22.2793278, 45.5202471, 41.8344086, 28.40037685, 14.5948914, 54.6870458, 42.1418541, 37.19420955, 51.0493286, 46.8183175, 41.8755616, -26.205, 7.0648306, 32.7762719, 30.544722, 20.9142238, 42.6502473, -0.2201641, 44.1767161, 20.2152358, 15.8693331, -12.0621065, 32.2228765, 49.1922443, 53.42968095, 42.4057219, 41.5510583, 37.808527350000006, 47.0245117, 23.135305, 27.3989266, 37.3886303, 10.2931062, 45.3969506, 12.1066695, 20.9876996, 31.2322758, 20.5927744, 44.9772995, 16.7967129, 44.9417468, 54.36119285, 56.9493977, 33.38034745, 47.6038321, 41.6509502, 48.4283182, 53.535411, 50.938361, 48.8588897, 50.8465573, 8.0610907, 19.0759899, 19.8887438, 40.7596198, 43.2073873, 13.3617562, 43.8519774, 52.4006632, 41.9960924, 41.3281482, -31.9558964, 40.640496, 48.1516988, -43.530955, 15.8798137, 43.2363924, -30.0324999, 51.3406321, -18.1415884, 10.2153093, 23.1301964, 14.9162811, 50.2137321, 35.6828387, 32.7174202, 11.568271, -0.62881505, -1.2832533, 11.083365350000001, 28.5421109, 20.6407176, 43.3211301, 51.772806599999996, 9.9051382, 50.1106444, 51.2254018, 50.6365654, 29.9759983, 57.1482429, 49.8955367, 43.6858146, 53.4071991, -8.4982478, 51.218194499999996, 58.3801207, -45.0321923, -25.2800459, 17.360589, -33.9617051, 34.7743988, 46.8836527, 54.9738474, 45.5548793, -25.7459277, 51.0460954, -2.8971725, 31.6258257, 25.6802019, 35.9127609, 52.131802, 34.6198813, 50.44876, 39.100105, 22.2973142, 36.0625843, 5.8247628, 20.0462328, 29.4246002, 38.1112268, 31.9515694, 6.112331, -34.9281805, 14.5589875, -8.6908357, 43.6044622, 5.9780066, 35.1799528, 26.2043691, 22.8015194, 45.3271752, 17.0, 1.5574127, 51.9244424, 43.1220119, 41.2230048, 23.7952809, 39.9622601, 12.9366674, 64.145981, 37.0162727, 43.541224400000004, -27.4689682, 33.88922645, 47.2186371, 55.6052931, 48.7784485, 16.4119945, 7.55984195, 44.3190159, 15.3503192, -25.1076701, 53.4794892, -8.6212371, 27.708317, 12.5672395, 44.1168594, 44.7720845, 49.1913033, 52.6361398, -16.9206657, 16.5735957, 36.550238, 43.6112422, 9.931308, 19.7073734, -34.9206797, 52.4796992, 34.6852901, 29.6519684, 20.6720375, 49.453872, 44.811349, 49.0521162, 39.7683331, 46.877229, 53.7435722, 18.4801972, -45.8740984, 12.3051828, 27.5236998, 30.0860459, 26.2122345, 40.4416941, 39.1014537, 33.6251241, 17.9712148, 12.2431693, 47.1615341, 19.7589517, 50.776351, -32.9595004, 39.613432, 22.68394485, -13.5170887, 37.270973, 35.4729886, 35.777103, 40.4864444, 55.17108, 30.3118769, 51.2211097, 45.94780155, 57.7072326, 33.4484367, 44.2337861, 38.3436365, 19.3580871, -12.9822499, 20.9670759, 44.648618, 37.9747645, 33.2336952, 35.2272086, 53.2190652, 48.23251, 25.9140256, 35.2072185, 32.3140354, 42.4415238, 37.5385087, 51.3582945, 44.841225, -54.806115899999995, 43.8975558, 34.7465071, 47.0708678, 39.7284945, 43.3435855, 42.498275, -16.3988667, -13.8343691, 30.6943566, 49.9655521, -3.7304512, 36.7213028, 42.8501191, 23.3700501, 37.0628317, 36.9863599, 29.7589382, 6.8137426, -7.2459717, 15.2241917, 50.9281717, 43.451291, 22.6203348, 22.5445741, -8.0584933, 36.0726355, 26.1947962, 35.3738712, 32.3669656, 47.3744489, 22.1757605, 49.0068705, 36.1556805, 34.6086854, 10.4195841, 35.7803977, 21.0294498, 36.3302284, 38.4224548, 41.6529143, 43.5460516, 26.3013982, 11.9477338, 39.9207886, 60.1674881, 45.7578137, 44.9391565, 46.097995, 38.7893909, 29.8988618, 31.8122623, -22.910832, 41.504879, 28.209538, 30.8327022, 35.9423833, 53.6466645, 32.4610708, 36.1622767, 50.013064, 43.0481221, 38.5810606, 41.1035786, 33.4709714, 32.3512601, 32.5010188, 51.47091115, 42.3315509, 11.9402416, 53.0758196, -6.9903988, 14.5567949, -10.3333333, 25.3356491, 8.5241122, 46.808327, 18.4724603, 28.8026443, 43.157285, 36.5436517, 37.8044557, 39.4667025, 47.5048851, 41.0799898, 30.1086537, 37.6922361, 6.3700386, 30.6598628, 38.7060457, 25.2653471, 48.7172272, 41.9818469, 36.06523, 0.5262455, -25.4295963, 25.4381302, 60.2047672, 36.5277607, -34.9058916, -42.8825088, 37.7742152, 36.7394421, 42.317099, -20.1624522, 48.2765903, 51.897077, 18.521428, 37.1641544, 22.33164705, 42.8620287, 43.6166163, 35.7688425, 33.5635206, 34.769536, -5.805398, 40.03813, 45.5031824, 34.0346534, 4.6533326, 22.9912348, 29.5541432, 6.9387469, 50.3219015, 42.444553, 42.0979745, 23.1146515, 25.7561647, 23.0205969, -22.5776104, 48.1113387, -1.45056, -8.334584, 44.40726, 9.9261153, 52.080985600000005, 2.3293744, 42.5814314, 43.7159395, 4.5986817, -34.4243941, 24.8843019, 51.0538286, 1.503555, 37.9577016, 40.6938609, 54.8982139, 39.08854049999999, 44.4938203, -5.1342962, 44.5126379, -33.4377756, 51.380926349999996, 34.5361067, 41.4086874, 45.7874957, -17.831773, 31.2237434, 30.3321838, 36.2516475, 35.2521996, 45.3573429, 42.8142432, 45.1875602, 34.6634288, 35.6892523, 0.3177137, 43.1579812, 30.247601, 4.8895453, 34.5793131, 32.0852997, 55.9533456, 43.0161693, -31.4173391, 12.054616800000002, 30.2265949, 37.104153, 52.0191005, -22.90556, 32.5007031, 39.063956, -8.8272699, -6.1753942, 40.1936892, -19.9227318, 34.851354, 9.3040623, 16.4639321, 52.6285694, 22.5816619, 46.2087066, 10.9428763, 32.844017, 26.1908241, 26.2235041, 6.3330586, 60.4517531, 52.23433665, 47.9078244, -18.9100122, 39.1670396, 19.439884550000002, 39.7686055, 35.8272257, -8.35760475, 44.7319094, 26.1223084, 39.9527237, 17.5211917, 40.335345, -4.2694407, 55.7121643, 53.3497645, 19.1943294, 36.1867442, 28.8638424, 26.1805978, 33.8439408, 31.3151708, 31.1171441, 35.5120831, 62.2393431, 31.098207, 55.8609825, 36.508976, 41.3082138, 39.94918695, 29.0948207, -37.9977225, -29.902631, 42.2885656, 34.0007493, 51.5142273, 35.1594647, 27.1752554, 41.6819935, 36.17262995, 40.3754434, 31.199004, 6.0328139, 7.3777462, 33.570499, 37.3361905, -1.950851, 51.1282205, 39.7589478, 33.5524785, 43.4652699, 28.4157769, -29.861825, 42.2799929, -17.333333, 33.5675527, 42.4184296, -20.4640173, -8.5837726, 20.28105765, 26.2159066, -20.2140657, 37.2633325, 25.702096, 43.074761, 40.1075089, 36.8968052, 52.1518157, 32.6405377, 35.6009498, 8.479004, 46.601557, 40.1827657, 22.8938884, 41.1257843, -11.6642316, 16.6835469, 38.2493581, -27.0988126, 65.0118734, 42.6638771, 22.0000001, 36.9903199, 40.8337963, 35.8713, 40.5863563, 51.7520131, 34.7494749, 35.3872218, 38.2542376, 26.2967719, -44.6941691, 42.8765615, 28.106471, 42.2681569, 52.5089759, -2.5295265, 40.6178915, 29.1871986, 44.0234387, 40.4233142, 15.3518378, 37.4946568, 52.3657098, 31.6343083, 27.9511188, 43.7698712, 30.1600827, 11.8948389, 40.6022059, 26.2037247, -23.1860383, 51.842574850000005, 40.4797828, 36.3134398, 41.6612561, -23.5003451, 8.034499199999999, 39.1404477, 9.0643305, 6.8393788, 10.5256264, 27.4989278, 46.8087794, -21.1343401, 21.19903535, 34.9236095, 23.6871297, 40.1164841, 47.41066, -16.4955455, 11.8763836, 37.7385507, 19.2716301, 29.03970805, 41.5910323, 30.5955289, 53.2744122, 39.2908816, -6.7137044, 41.7571701, 40.5092961, -17.3936114, 38.8339578, 41.4996562, 7.7382813, 42.1294712, 33.8829996, 43.8130735, 34.9531295, 33.9566098, 36.3967623, 33.1154683, 33.9533546, 22.9372087, 3.5896654, -16.680882, 34.0747444, 7.2930922, 53.7974185, 46.2062319, 43.17389195, 49.694285, 30.2240897, 34.5978596, 30.4211144, 16.5087586, -0.19299, 47.7981346, 37.6390972, 43.1009031, 35.021041, 37.7986184, 44.2613967, -15.4164124, 57.0462626, 41.8239891, 30.6743643, 34.3952765, 11.10684475, -4.05052, 17.7231276, 39.7990175, 32.182598, 32.5369356, 13.4410165, 38.6264256, 11.2450558, 40.8358846, 6.487679099999999, 23.356377, 50.9025349, 40.6549157, 42.9995032, 34.4263886, 21.2379469, 13.4727446, 24.8755492, 47.0451022, 27.9946969, 37.4137536, 4.7676576, 39.8454163, 31.3869323, 11.127834, 22.5414185, 42.9632405, 50.8036831, -16.5048228, 41.5235808, 46.1030443, 49.163877, 49.7596208, 48.7544012, 28.3754144, 37.08418, 28.6581723, 8.8879509, 41.9758872, 38.7566055, 18.833333, 45.6496485, 34.729847, 20.0112475, 21.2321047, 46.2017559, 42.5069391, 9.5170602, 35.1748976, 28.4963633, 29.6060218, 35.9556924, 23.2790444, 28.2996183, -20.3200917, 52.2055314, 32.0615652, 43.5854823, 34.022405, -6.9344694, 22.5213807, 45.5616075, 52.9534193, 52.5150949, 27.7237996, 37.870009, 36.0638034, 42.7337712, 48.584614, 29.5647398, -7.1215981, 44.3893113, 32.2990021, -6.1258942, 42.2713945, 36.7753606, 43.8014053, 43.3709703, 39.906217, 44.4761601, 7.1172633, 53.9590555, 5.947822, 41.3123363, 40.9965446, 52.3744779, 29.3093002, 32.5221828, 52.0799838, 37.8725487, 22.5000001, 51.5887845, 1.2146013, 10.6572678, 25.6093239, -21.7558619, 37.1734995, 36.3320197, 35.7032751, 59.9133301, 45.806691349999994, 33.7475203, 52.4081812, 51.456659, 45.92956455, 20.4317585, 31.549333, 50.7194784, 56.7291997, 31.666960500000002, 50.879202, -2.2895167, -23.6463741, 38.9181714, 35.6876096, 49.4093582, 49.872775, 13.6989939, 36.8529841, 10.50614075, 29.8622194, 31.3016935, 35.1851045, 24.5164639, 27.7951482, 21.19177615, -23.1887668, 66.4976214, 12.5268736, -8.5536809, -29.310054, 40.2337289, 42.6511674, 23.1102636, 36.4072377, 28.6517178, 19.292545, 55.7127529, 14.6222328, 13.0836939, 47.9184676, -18.9188041, 38.4493315, 42.8184538, 12.3681873, 40.8228016, -34.787093049999996, -8.8285734, 42.03726, 39.1649125, 45.4077172, 29.000652549999998, 18.465299, 6.130419, 11.0018115, -2.9888297, 51.8784385, 6.698081, 47.2654296, 42.3192548, 31.8228094, 21.9812746, -6.3718272, -12.6076318, 48.0517429, -24.6581357, -5.8951725, 51.3154546, 25.238704, 31.5841689, 36.77101545, 38.246242, -4.1030531, 36.2974945, 42.6829765, 47.6571104, 42.9956397, 24.163162, 43.3224219, 31.4933074, 33.9137085, 10.9810761, 8.7342288, 22.273734, 32.5866826, 30.4459596, 50.8389149, 22.7203616, 31.8782976, 51.585618749999995, 39.049011, 35.0457219, 41.6749893, 24.4480637, -3.1316333, 5.320357, 21.2737079, 41.1670412, 43.4887907, 41.8933203, 46.9482713, 19.7540045, 19.1214072, 22.3053263, 50.7255794, 35.1987522, 41.8692972, 34.7333334, 29.2108147, 48.7081906, 41.5263603, -8.6831728, 36.617331, -33.0458456, 32.8795022, 31.9973662, 31.7819242, 30.0443879, 29.9992425, 25.4557366, -17.853135, 41.139981, 36.0127749, 41.764582, -26.2227778, 34.4221319, 39.9395724, 44.5384728, 31.5118281, 56.1496278, 47.3523672, 35.444991, 35.7947315, 4.8472017, 51.984257, 38.2677554, 32.2352666, 35.124945, 6.1461619, 4.0429408, 26.6059432, 45.8992348, 51.8995685, 27.2939333, 39.0167979, 47.3215806, 38.2544472, 60.3943055, 40.8088861, 37.9317766, 33.4942189, 38.0739964, 13.519054, 33.9814746, 27.7809047, 43.9470982, 26.640628, 40.4906216, 33.577752399999994, 26.8381, 26.5914524, -8.7724902, 52.2695736, 54.596391, 35.48001575, 38.8462236, -33.0244535, 41.0534302, 45.7988273, 64.175029, 39.1723908, 45.9917589, 61.4980214, 19.29348285, 32.896672, 43.9492493, 47.6571934, 11.2320988, 42.4966815, 32.665135, 27.833333, 50.1171903, 34.6553944, 43.0349931, 42.7656251, 33.3864655, -13.9875107, 49.8879177, 63.4305658, -28.0023731, 23.2584857, 36.19924, -35.2975906, 10.506098, 24.3284746, 30.4380832, 36.0620781, 38.1301968, 13.524834, 49.5978157, 11.1427624, 4.1779879, 29.9870753, -51.69487225, 20.8689257, 43.2961743, 43.061936, 10.2375144, 24.4801069, 30.2489634, 41.2587459, 51.2085526, 17.9640988, 0.5184021, 6.809107, 53.0775961, 45.6794293, 28.5811076, 42.5006243, 39.5261206, 32.6248622, 51.4816546, 40.0149856, 3.8689867, -15.9277296, 31.9827896, -6.8160837, -7.9771206, 27.9963899, 14.7421699, -8.6524973, 18.062922200000003, 35.3676472, 18.547327, 26.1003392, 24.7807798, 39.579155, 24.4538352, 21.1498134, 26.9154576, 44.0505054, 22.050466550000003, 45.4371908, 32.8406946, 51.4392648, 17.7395201, 34.3916058, 69.649208, 42.8867166, 40.6460921, 22.8193063, 10.6498095, 42.2625621, 30.8942995, 49.4404591, 9.0107934, 35.0841034, 37.0664467, 40.4590505, -25.966213, 38.4404925, 35.5570501, 43.5515198, 45.4641943, 10.3825318, 27.3365805, 34.6422939, 23.0216238, 45.4871723, 37.042257, 36.4071112, 13.0977832, -21.2074736, 34.3430507, 32.6200973, 34.4838502, 31.8457149, 18.556551, 35.1343368, 52.707755, 29.299328, 26.4609135, 33.4946353, 40.0919817, 52.1067449, 31.3536127, 38.0464066, 17.1527588, -7.9778383999999996, 31.9030543, 28.687547, -7.5692462, 18.9194213, -34.9632342, 38.9719384, 33.6306299, 32.7876012, 28.42826235, -8.5199633, 53.3806626, 44.0581728, -16.3332828, 23.1608938, 15.3389667, 26.7300936, -9.437797549999999, 33.1580933, 39.7034723, 37.4443293, 31.263235, 30.5951051, 12.61326555, 34.6932379, 33.4270275, 43.1672254, -28.6483333, 35.1433533, 30.9090157, 43.7009358, 33.6170092, 40.7896239, 61.2163129, 26.715364, 25.1737019, 39.2169525, 34.1976308, 39.0856735, 32.44645, 37.3893889, 41.1075276, 35.9170217, 40.0940758, 40.34833555, 17.1184569, 15.35, 40.63367, 40.924271, 24.5625566, 42.3636331, 37.5120997, 39.7459468, 30.421309, 38.4852037, 46.7729322, 39.415779, 37.4334738, 33.9516458, 38.974390099999994, 35.1335022, 34.2066591, 36.140807, 23.7861979, 44.0869329, 40.5508527, 45.4384958, 46.5927425, -4.3217055, 43.6610277, 33.0010496, 34.7487244, 43.2630018, 17.5000543, 37.4519671, 42.437072, 36.2231079, 32.0809263, 24.8546842, 19.877263, 36.343694049999996, 16.0000778, 11.2536716, 33.1958696, 33.1141426, 5.5571096, 33.2000368, 30.7334421, 44.5645659, 62.76712545, 37.496904, 49.1856637, 35.1139509, 26.4367824, 43.73844905, 4.6373012, 39.4862777, 37.9716991, 47.5581077, 29.101628050000002, -8.7278354, -12.46044, 24.471153, 43.4832523, 52.383705750000004, 34.6196539, 51.1777781, 25.0783456, 33.743764999999996, 30.1957677, 31.79592425, 49.6112768, 38.5425835, 44.5263107, 38.2325829, 53.2005936, 30.7474425, 39.5404679, 0.390002, 18.4257128, 26.12744855, 26.339806, 39.5182427, 43.419754, 46.8701049, 39.2501246, 38.0429742, 25.2856329, 35.9603948, 3.4517923, -17.7414972, 35.9957253, 27.4989086, 36.6507007, 38.2971367, 44.9391565, 32.010998, 45.639544, 6.5269033, 19.30037305, 44.766764249999994, 10.4541312, 38.8579735, 32.555664, 44.0206919, 2.0349312, 41.9602718, 25.3575436, 52.3713394, 58.42673005, 30.3674198, 36.1911624, 39.915463, 45.0677551, 33.2095614, 32.2956076, 18.0250713, 33.3024309, 32.6707877, 44.017763900000006, 46.5218269, 31.5656822, 23.61515, 36.7071591, 31.4649685, 46.0050102, 33.822986549999996, 24.638916, 33.74079965, 11.0248119, 18.341137, 10.6333401, 38.50935, 36.364519, 9.50139445, 19.7490157, 30.694088, 36.1023552, 46.6855231, 31.2457442, 32.0898005, 36.3496, 47.4250593, 33.6038766, 18.3403061, 38.7367448, 14.693425, 34.0123846, 27.8328352, 32.8191218, 39.6295951, 29.394644, 37.451434500000005, 42.5846773, 34.5260109, 36.819626, 12.1459907, 12.4345376, 36.583333, -54.283545, 33.6938118, 9.3348046, 37.9404648, 10.302834, 31.4220558, 30.1979793, 32.1288331, 14.1057433, 32.1525312, 10.0844025, 47.6144219, 41.903411, 30.4952371, 37.461928, 32.4935378, 4.9371143, 14.6367927, 15.5062156, 12.1472869, -9.4743301, 14.8954912, 41.9263991, 31.3231643, 15.5635972, 29.3797091, 34.7109786, 33.5130695, 11.8145966, 28.1450774, 4.3907153, 16.3217674, 32.3969935, 21.5810088, 21.420847, 6.328034, -22.2745264, 12.1191543, 50.4500336, 55.7504461, 49.841952, 53.9024716, 46.4843023, 48.7481718, 59.938732, 54.7261409, 49.9923181, 45.7684014, 63.3233807, 48.4680221, 47.9102734, 56.866557, 55.1598408, 51.530018, 54.991375, 56.839104, 54.710128, 54.3107593, 55.0951738, 47.2610441, 53.198627, 55.0282171, 58.324653850000004, 43.17199225, 55.7823547, 56.3264816]   
df['Longitude']=[-9.399580512067734, -9.1365919, -80.19362, -15.6062305376665, 19.0403594, -97.7436995, -8.6107884, -58.4370894, 44.8014495, 98.9858802, 23.3221789, 21.0067249, 100.4934734, 20.4568974, 28.9662187, 98.3529292, -16.981752048847362, 100.0305710845208, 108.212, -99.1331785, 14.4212535, 18.417396, 126.9782914, 103.8194992, -79.3839347, 21.2257474, 13.3888599, 18.0710935, -0.3763353, 16.4399659, 101.6942371, -46.6333824, 112.46309700455788, 19.9368564, 77.590082, 100.2559077, -84.3902644, 16.97819633051261, -87.0779503, 15.9771774, 26.1027202, 44.5126174, 30.6954846, 144.9631608, -157.717299, 174.7631803, 23.5899542, -16.621447121144122, 151.2082848, 14.5069289, -48.5496098, -123.113952, -3.7035825, -75.573553, 121.5636796, 19.8451756, -0.1276474, -115.148516, 25.6102746, -118.242766, 115.2622931, 11.5753822, 12.5700724, 23.7283052, 2.1774322, -43.2093727, -8.702467283736809, -74.0060152, 24.7453688, 106.6983801, -104.9848623, 16.3725042, -86.8510468, -77.0365427, -71.0582912, 18.7710437, -122.419906, 174.7772114, 4.8936041, -82.458444, 22.9352716, -8.4294632, 10.000654, 114.1628131, -122.6741949, 23.4841698, -14.004869713542156, 120.9782618, 25.2829111, 24.7499297, -8.585577395168343, 13.7381437, -71.2143518, -87.6244212, 28.049722, 125.6080623, -96.7968559, -9.709128, -100.743784, 18.0924947, -78.5123274, 28.6507598, -87.42953531439221, -97.0726428, -77.0365256, -110.9748477, 16.6113382, 14.592912670292268, -82.1853837, -8.4280045, -25.473085286479595, 28.8322923, -82.3589631, 33.6741556, -5.9953403, 123.9020773, -75.6851593, -68.9351307, -101.0, 121.4692071, -100.3902245, -93.2654692, 96.1609916, 26.0236504, 18.62860883362069, 24.1051846, 126.54825445912047, -122.3300624, 41.6360085, -123.3649533, -113.507996, 6.959974, 2.3200410217200766, 4.351697, 98.9161607, 72.8773928, 102.135898, -111.8867975, 27.9166653, 103.8590321, 18.3866868, 16.91973259178088, 21.4316495, 19.8184435, 115.8605801, -8.6537841, 17.1093063, 172.6366455, 108.3273939, 76.9457275, -51.2303767, 12.3747329, 178.4421662, 103.9880443, 113.2592945, -23.5095095, 19.00588775660632, 139.7594549, -117.1627728, 104.9224426, -90.3638752022324, 36.8172449, 120.91639641658887, -81.3790304, -105.220306, 21.8959232, 19.47849495849145, 126.05118878624401, 8.6820917, 6.7763137, 3.0635282, -90.0782127, -2.0928095, -97.1384584, -79.7599337, -2.99168, 116.0521267, 22.554677562145155, 26.72245, 168.661, -57.6343814, 78.4740613, 25.6207519, 32.4231586, 17.846663499449168, -1.6131572, 18.6953685, 28.1879101, -114.065465, -79.0043742, -7.9891608, -100.315258, 14.5020708, -106.660767, 135.490357, -104.61731, -94.5781416, 73.1942567, -94.1574328, -55.1703941, 110.1956502, -98.4951405, 13.3524434, 35.9239625, 125.1717145, 138.5999312, -90.7333258, 115.1641975, 1.4442469, 116.0728988, 129.0752365, -98.230082, 86.2029579, 14.4412309, -96.5, 110.3439862, 4.47775, -123.4131383, -111.9738429, 86.4309638, -83.0007065, 100.8864587, -21.9422367, -7.9351771, -96.73366465001021, 153.0234991, 35.50255852895232, -1.5541362, 13.0001566, 9.1800132, 120.5933729, 99.05728875374527, 23.7965614, 74.10178169795284, -64.3494964, -2.2451148, 115.126168, 85.3205817, 99.954586, 15.2353257, 17.1917651, -122.8491439, -1.1330789, 145.7721854, 120.4089899, -82.5594293, 3.8767337, 76.2674136, -155.08158, -57.9537638, -1.9026911, 33.0332657, -82.3249846, -103.338396, 11.077298, -91.4984941, -122.329479, -86.1583502, -96.789821, -0.3394758, -69.942111, 170.5035755, 76.6553609, -99.497352, -94.1018461, 127.6791452, -79.9900861, -84.5124602, 130.6180016, -76.7928128, 109.1898675, 27.5836142, 99.6734592, 6.083862, -60.6615415, 2.880353635614984, 91.78913897849154, -71.9785356, -79.9414313, -97.5170536, -5.803792, -86.1336351, -118.7949873, -95.4560512, 4.3997081, -66.65347069241668, 11.9670171, -112.074141, 28.6255665, -0.4881708, 98.439839, -38.4812772, -89.6237402, -63.5859487, -87.5558483, -8.500424800741666, -80.8430827, 6.5680077, -101.296273, -97.4890856, -101.8338246, -106.7798078, 19.2621081, -77.43428, 7.473296, -0.5800364, -68.3184972880496, -78.8635324, -92.2896267, 15.4382786, -121.8374777, 17.8076584, -92.3328743, -71.5369607, -171.7692793, -88.0430541, 16.9705651, -38.5217989, -4.4216366, -106.325138, 85.3250387, 37.3792617, 35.3252861, -95.3676974, -58.1624465, 112.7378266, 101.9313754, 11.5879359, -80.4927815, 120.3120375, 114.0545429, -34.8848193, -79.7919754, -98.1836216, -119.0194639, -86.3006485, 8.5410422, 113.5514142, 8.4034195, -95.9929113, -98.3903305, -75.5271224, -78.6390989, 105.8544441, -119.2920585, 27.1310699, -83.5378173, -80.2493276, -98.1624501, 121.9392632, 32.8540482, 24.9427473, 4.8320114, -123.033121, -64.80011, 0.1633256, -93.9288723, 119.9691539, -68.2001376, -73.9696822, 83.991402, -83.2784851, -77.7956949, -1.7822482, -84.9880449, -86.7742984, 22.016143457294312, -76.1474244, -121.493895, -80.6520161, -81.9748429, -95.3010624, -116.964662, -2.5790334025708415, -83.0466403, 108.4375758, 8.8071646, 110.4229104, 121.0211226, -53.2, 83.0076292, 76.9360573, -100.783739, -77.9217357, -96.9766308, -77.615214, 32.0019211, -122.2713563, -87.4139119, -111.29189, -85.1386015, 78.2916193, -97.3375448, 99.7928634, 104.0633717, -88.0949352, 55.2924914, 21.2496774, 44.1117811, -119.0167679, 101.4515727, -49.2712724, 81.8338005, 24.6568435, -87.3588703, -56.1913095, 147.3281233, -87.1133304, -119.7848307, -83.0353434, 57.5028044, -116.553247, -8.4654674, 73.8544541, -120.7678602, 103.85753892309174, -112.450627, -116.200886, -119.2470536, -101.879336, -92.2670941, -35.2080905, -76.3056686, -73.5698065, -5.0161926, -74.083652, 120.184982, -81.2207673, 79.8541134, 11.9178807, 27.208264534658166, 19.0954528, 114.4107202, 112.72964, 113.7457788, 17.0772739, -1.6800198, -48.4682453, 115.6409359, 8.9338624, 78.1140983, 5.12768396945229, 102.2880962, 18.2177461, 10.4018624, 101.0900236, 150.89385, 102.8324209, 3.7250121, 103.7495586, -121.2907796, -89.5891008, 23.9044817, -120.05035277301698, 11.3426327, 119.4124282, -88.0125794, -70.6504502, -2.37418971512966, -117.2911565, -75.6621294, -108.49607, 31.045686, -85.3933906, -81.655651, -118.852583, -92.6888936, -122.6067583, -73.9395687, 5.7357819, 135.5032721, 51.3896004, 32.5813539, -79.2441003, 121.31938451585901, 114.9417574, -118.1171108, 34.7818064, -3.1883749, -83.6900211, -64.183319, 102.33734259872213, -93.2173759, -113.5841313, 8.531007, -47.06083, -94.74049, -108.5507317, 13.2439512, 106.827183, -85.3865271, -43.9450948, -82.3984882, -82.12848154978464, 107.5863388, 1.2923049, 113.0761073, -119.119919, 108.2830253, -97.1430671, -97.6959794, 50.5822436, 5.6221058, 22.2670522, -0.9028072768185829, -97.0592028, 47.5255809, -86.5342881, 72.8803834000775, -94.8466322, -90.6949871, 116.08115279184318, -93.21772, -80.1433786, -75.1635262, 102.6680012, -75.9279495, 15.2712256, 12.563150889107241, -6.2602732, 72.9701779, -94.1288142, 78.80577833091104, 91.753943, 9.400138, -85.8552161, -97.727796, 24.0191544, 25.7458236, -97.3427847, -4.2488787, -4.88562, -72.9250518, 4.0529515568716175, -110.9692202, -57.5482703, -71.2520253, 18.8419505, -81.0343313, 7.4652789, 126.8515034, 78.0098161, -85.9766671, 27.919418145633188, 49.8326748, 29.894378, 80.214955, 3.8972497, -86.765783, -121.890583, 30.061507, 71.4306682, -84.1916069, 119.1075773, -80.5222961, 79.45264865437258, 31.009909, 118.9426131, -61.5, -7.618756606822727, -71.1061639, -54.6162947, 116.10685, 85.81140263163533, -98.3252932, -70.1524646, 127.0287472, 32.647186, -89.3837613, -89.3768539, -76.2602336, 4.4811088666204295, -85.4626848, -82.5540161, -13.26795, -120.510842, 29.0677305, -109.9200604, 16.8620293, 27.4826264, -92.61551841991157, -122.039966, -48.64454838500228, 25.4716809, 21.1640849, -102.5000001, -86.4436018, 111.6730788, 128.6018, -122.391675, -1.2578499, -77.4208221, -94.4248983, -85.759407, 73.0351433, 169.1364637, 74.6070079, -80.6371513, -83.7312291, 6.0943765, -44.2963942, -75.3786521, -82.1400923, -92.4630182, -104.7091322, 75.1379848, -120.8465941, 5.2210432, 74.8736788, 116.3530413, 11.2555757, -85.6545729, 8.5364136, -75.4712794, 78.1573628, -45.8866816, 5.838960628748229, -88.9939147, -82.3534728, -91.5299106, -47.4582864, 98.81788050252146, -121.6169108, 7.4892974, 81.8327629, 76.2132542, -82.5748194, 130.3734556, -175.201808, 81.3979545573657, 33.6236184, 86.9746587, -88.2430932, 40.101986, -68.1336229, 75.3737973, -121.4201395, 73.23586552517415, -13.636291608604138, -93.6046655, -96.3071042, -9.0490601, -76.610759, 108.5608483, -88.3147539, -88.9843937, -66.1568983, -104.825348, -81.6936813, 98.7709086, -80.0852695, 130.8749015, 125.3171216, -120.4358577, 116.7928916, -97.8791341, 129.7874339, -117.3961623, 89.2852741, 98.6738261, -49.2532691, 74.8204443, 80.6350768, -1.5437941, -90.6802594, 16.55652349457852, -112.851562, -92.0198427, 119.215813, -9.5830626, 80.6185102, -79.272301, 13.0464806, -120.9968782, -75.2326641, 135.7556075, -121.2117148, -88.4069744, 28.2824793, 9.9215263, -71.4128343, -96.3699632, 129.31544896587815, 76.10995510466662, 39.667169, 83.3012842, -89.6439575, -95.7893178, -93.6952689, -16.56275092072591, -90.1995853, 75.7754716, 14.2487679, 99.3024780130819, 116.6775918, -1.404189, 109.8344586, 125.9816054, -117.3008784, 81.6336833, 144.75201822544284, 118.6723605, -122.8950075, -80.6366144, -79.1422464, 7.0188527, -88.9524151, 120.9764683, 77.3415747, 88.35769124388872, -85.6678639, -1.075614, -68.1624337, -90.5770967, 11.129840331695068, -123.938122, 6.6441878, -122.478836, 121.3812013, -94.51323, 121.4163876, 76.5955013, -91.6704053, -90.42352, -98.0, 13.7772781, -86.5859011, 73.7902364, -86.73222812171971, 6.1466014, 1.5212467, -13.6998434, 33.3638568, 34.5145652, 52.5378041, -80.0053176, 116.0876485, -82.4522702, -40.3376682, 0.1186637, 118.7915079, 39.723109, -6.834543, 107.6049539, 113.3656141, -94.1642004, -1.1496461, 5.4768915, 106.9191408, 112.5452739, 120.3781372, -84.5553805, 7.7507127, 106.5478767, -34.882028, -79.6901736, -90.1847691, 23.5998111, -89.093966, 3.0601882, -91.2395429, -8.3959425, 116.3912757, -73.212906, -73.1156668, -1.0815361, 80.5482919, 69.2787079, 15.1405690365004, 9.7385532, 120.0703805, -93.7651944, 4.3113461, 32.4942749, -100.4949145, 4.7760237, -77.2782755, -61.5180173, 85.1235252, -41.3326895, -3.5995337, -94.1185366, -0.6492976, 10.7389701, 12.206315763116372, -116.9719684, -1.510477, -0.9696512, 6.934005265146571, -86.92027456597663, -97.1466695, -1.8767712, -111.3885221, 120.79080112776134, 4.7011675, -80.0983887437901, -70.3980033, 121.6282945, -105.9384569, 8.694724, 8.651177, -89.1914249, -75.9774183, 103.26205644863938, 121.6203873, 120.5810725, 136.8998438, 117.6424895, -97.3942053, 72.9544157804841, -46.884506, 25.7192101, -70.0356845, 125.5784093, 27.478222, -111.6587085, -73.754968, 112.99801241836545, -105.573284, 77.2219388, -99.6569007, 21.1350469, -90.5185188, 80.270186, 106.9177016, -48.2767837, -78.8688833, -1.6442556, -1.5270944, 114.880567, -68.43818677312292, 115.087937, -88.2810994, -119.7666496, 11.8734455, 77.76822875372174, -66.116666, 1.215829, 76.9628425, 104.756857, -0.4152837, -1.6230404, 11.3927685, -85.1824269, 117.2218033, 96.082375, -77.8093378, 15.7411039, -122.1768209, 25.9088474, 22.408589, 9.4924096, 110.1750728, 130.543387, 3.509589562138503, 21.7350847, -81.0481341, 59.6059232, -89.0226793, -117.2613936, -71.4547891, 120.6478282, -1.9838889, 120.3074357, -98.4933873, -74.7988827, 76.7257086, 113.5721327, 117.0130019, -91.18738, -0.12693017575934035, 75.8681996, 120.5514595, 5.066061908648313, -95.677556, -85.3094883, 123.458674, 118.3856331, -59.9825041, -4.016107, 110.3548318, -73.2048348, -112.03628, 12.4829321, 7.4514512, 96.1344976, -70.6432141, 70.8028377, -3.5269497, -111.6518229, 123.924157, 36.7166667, -81.0228331, 44.5153353, -88.0840212, 115.2564626, 101.7761976, -71.6196749, -111.757352, -102.0779482, -95.6258199, 31.2357257, 120.576854, 119.0029034, 25.861429, -104.820246, -5.6048873, -72.6908547, 27.89, -119.7026673, 119.5894129, 10.935960870530739, -9.7620903, 10.2134046, 123.9164653, 139.636768, -107.21208446076524, 31.5951655, 5.9108573, 140.8691498, 76.326238, 7.901173504229375, 6.8019088, 9.706203, -81.9806771, 6.1288847, -2.0711559, -80.3503283, 125.7473609, 5.0414701, -104.609141, 5.3259192, -96.7077751, -121.696026, -111.9260184, 46.2961952, 144.83783877764978, 134.3649289, 120.6512652, 12.630768610516004, -81.8723084, -98.9472344, -117.18845420205307, 80.9346001, 127.9773062, 115.1723944, 6.236339604766698, -5.9301829, 44.360072742328676, -77.3063733, -71.5517636, -73.5387341, 126.5303997, -51.7355386, 127.4285128, 9.264881035005711, 23.7603118, -100.46459244066614, 13.1777923, 4.8059012, -117.4235106, -74.1950883, -96.4058782, -114.47603157249804, 78.166667, -122.9543022, 133.9194595, -87.922497, -71.4677032, -86.78247922885197, 33.768144, -119.4959025, 10.3951929, 153.4145987, 77.401989, 37.1637253, 149.1012676, -66.9146017, 109.4239676, -84.2809332, 103.8318566, -121.2724473, 2.109823, 11.0038064, -74.1169296, 73.5107387, 31.2118063, -57.84522966123792, -105.4407022, 5.3699525, 141.3542924, -67.5890216, 118.0853479, 120.2052342, -95.9383758, 3.226772, 102.6133707, 25.2057292, -5.273263, 4.9057311371764385, -111.044047, 77.3031364, -90.6647985, -119.8126581, -115.448325, -3.1791934, -105.270545, 11.5213344, -5.7160872, 120.8904588, 39.2803583, 112.6340291, 120.695345, -91.1559757, 115.2191175, -15.962904570040749, 24.4736079, -72.3395928, -80.399513, 110.4922502, -104.8769227, 54.3774014, 79.0820556, 75.8189817, -123.0950506, -159.55876775017003, 12.3345898, -83.6324022, 5.478633, -88.03365155356477, 132.4518156, 18.9543434, -78.8783922, -111.4979963, 108.3627211, -71.6443596, -71.8018877, 120.0851114, 1.0939658, 38.7612525, -106.650985, 27.484418184610497, 128.9038646, 32.56745, -122.7141049, 45.4425963, 7.0134418, 9.1896346, 7.8533226, -82.5308545, 50.8801184, 72.5797068, -122.80378, 43.0953954, 25.456663652707153, -59.6184184, -159.7708145, 108.9350839, -83.606568, -114.3224495, -102.367687, -68.3691611, 36.7496276, -2.7540658, -94.7945882, 80.3217588, -117.1473661, 113.2951122, 4.2736937, 118.4277477, -84.4970393, -89.0761718, 110.36722565020224, 35.1952255, 115.8540042, 110.828448, 102.4477333, -54.9439981, -95.2359496, 114.6423617, -79.9402728, 77.00270014657752, 179.1982548, -1.4702278, -121.3153096, -48.9525756, 79.9497702, 38.9326763, -12.9766773, 159.9624174786047, -117.3505966, 125.8888149, -122.1598465, 32.305505, 114.2999353, -7.984739136241295, 135.1943764, -117.6124179, 11.467181161339118, 153.6177778, 110.769063, 75.851601, 7.2683912, -117.9294401, -73.9598939, -149.894852, -80.0532942, 75.8574194, 9.1128513, -119.1803818, 117.1951073, -99.7475905, -122.0832101, 122.9902689, 139.4886305, 124.3977559, -74.6684230570877, -61.8448509, 44.2, 14.6026095, -98.338685, -81.7724368, -87.8447938, 122.1152599, -75.546589, -87.2169149, 106.2261926, -92.1251218, -77.4127562, 118.6689349, 131.2468368, 1.4197463178515335, -89.9668758, 117.278286, -5.3541295, 90.4026151, -103.2274481, -105.0668085, 10.9924122, -112.036277, 15.3125974, -70.2548596, 112.5292693, 113.6191483, -2.9350039, -88.2003115, -122.1779927, -76.54831458600927, -121.3877428, -81.0911768, 67.0207055, 75.3390241, 43.099715597302286, -61.7333373, -85.8709492, -117.3794834, -86.8473788, -0.2012376, -117.2425355, 76.7797143, -123.2620435, 6.168828066313526, -122.3330573, -2.1102277, 117.1600052, 50.1039991, 7.424224092532953, -75.569799, -75.0254256, 126.5527573, 7.5878261, -81.01059617608054, 115.1730049, 130.8410469, 39.6111216, -1.5592776, 4.643559696559635, 112.4477046, -115.5682504, -77.3383331, -116.22331524966393, 67.0172447, 35.21198075969497, 6.129799, 68.81521423589763, -109.056392, -122.636465, 5.7918548, 120.7510971, -119.7487235, 9.454001, -64.6232227, -80.25779068328761, -81.7786972, -119.988811, 87.319461, -113.995267, -119.951909, 114.5088385, 51.5264162, -83.9210261, -76.5324943, 168.3150163, 119.4041185, 113.1529127, 117.1140042, -122.2855293, -123.033121, 112.1163785, 9.278869369225038, 3.5774005, -81.37009366377936, 11.827938995640167, -84.0096613, 115.490696, -95.8640507, -88.5408574, 45.3419183, 129.96778069176605, 55.3882729, 4.5307624, 8.94694911492874, -89.0928155, 44.0094652, 127.5439794, 7.6824892, -87.5675258, -64.7827048, -63.0483073, 44.3787992, 51.6650002, 10.454430026192636, 6.6327025, 74.3141829, 58.5912467, 119.1557441, -100.4405094, 8.9520281, -116.46687017113196, 46.7160104, -116.36936983394352, 7.651853311586596, -64.932789, -85.4362722, 125.7555447, 6.60826, 99.99561927320003, -70.4128435, 111.2803074, 114.33633, 7.8585139, 34.7925181, 72.6782574, 127.3848, 9.3765878, 73.048136, -88.49537037472777, 125.4115285, -17.447938, 71.5787458, 112.9388078, 34.9983856, 118.1738712, 71.6638747, 25.392337493319413, -87.8212263, 69.1776838, 117.980675, -86.2746665, -86.8773498, 114.483333, -36.4946355, 73.0651511, -83.5720475, 58.3823487, -85.8385625, 73.0923253, 71.4724978, 20.0817204, -87.2040052, 74.1933745, 99.8258031, -122.192337, 12.4528527, 47.8090981, 121.4425255, 74.5411575, -52.3258307, -61.01582685063731, -88.0248937, -68.2740206, 147.1599504, -88.92572717284826, 8.7376029, 48.6792538, 32.5349123, 47.9735629, 137.7259431, 36.3095814, 42.8453061, 113.2384362, 18.5509126, -86.17868327530661, 119.4077008, 39.1653612, 39.826869, -10.797788, 166.442419, 15.0502758, 30.5241361, 37.6174943, 24.0315921, 27.5618225, 30.7322878, 24.5207477, 30.316229, 55.947499, 36.2310146, 39.0261044, 97.0979974, 35.0417711, 33.3917703, 53.2094166, 61.4025547, 46.034683, 73.371529, 60.60825, 20.5105838, 48.3642771, 36.611913, 39.72490856518773, 50.113987, 82.9234509, 81.52937237232149, 131.96425966543694, 49.1242266, 44.0051395]

df

df.to_excel('./7_final_info.xlsx')
df.to_csv('./7_final_info.csv')
